fast boot aware service
